requirements.txt
python-dotenv
sqlalchemy
pyopenssl
alembic
pycryptodome